Move collateral tests to separate modules
Move the collateral tests out of Test.Cardano.Api.Transaction.Autobalance
and Test.Cardano.Api.Experimental.Fee into new modules, split by API
flavor: Test.Cardano.Api.Transaction.Collateral for the traditional API
and Test.Cardano.Api.Experimental.Collateral for the experimental API.
The utilities shared by the two traditional-API test modules move to
Test.Cardano.Api.Transaction.Utils. The tests themselves are unchanged.

Suggested by @carbolymer in:
https://github.com/IntersectMBO/cardano-api/pull/1265#pullrequestreview-4774208788
