+++ cardano-api/cardano-api.cabal
@@ -408,6 +408,7 @@ test-suite cardano-api-test
     Test.Cardano.Api.EpochLeadership
     Test.Cardano.Api.Eras
     Test.Cardano.Api.Experimental
+    Test.Cardano.Api.Experimental.Collateral
     Test.Cardano.Api.Experimental.Fee
     Test.Cardano.Api.Genesis
     Test.Cardano.Api.GovAnchorValidation
@@ -422,6 +423,8 @@ test-suite cardano-api-test
     Test.Cardano.Api.RawBytes
     Test.Cardano.Api.Transaction.Autobalance
     Test.Cardano.Api.Transaction.Body.Plutus.Scripts
+    Test.Cardano.Api.Transaction.Collateral
+    Test.Cardano.Api.Transaction.Utils
     Test.Cardano.Api.TxBody
     Test.Cardano.Api.Value
